+++ packages/apple-targets/src/icon/with-image-asset.ts
