+++ packages/apple-targets/src/with-widget.ts
